update code snippets, add includes and new sections
