+++ examples/dotnet/DataBinding/MainPage.xaml
@@ -34,7 +34,17 @@
             </ListView.ItemTemplate>
         </ListView>
         <!-- :snippet-end: -->
-
+        <!-- :snippet-start:bind-embedded-xaml -->
+        <ListView x:Name="listItems" ItemsSource="{Binding IncompleteItems}">
+           <ListView.ItemTemplate>  
+                <DataTemplate>  
+                    <ViewCell>
+                        <Label Text="{Binding Summary}"/>
+                    </ViewCell>
+                </DataTemplate>
+            </ListView.ItemTemplate>
+        </ListView>
+        <!-- :snippet-end: -->
     </StackLayout>
 
 </ContentPage>
