+++ examples/dotnet/DataBinding/Models.cs
@@ -5,6 +5,7 @@ using Realms;
 
 namespace DataBinding
 {
+    //:snippet-start:binding-classes
     public class Employee : RealmObject
     {
         [PrimaryKey]
@@ -27,4 +28,24 @@ namespace DataBinding
             this.Items = new List<Item>();
         }
     }
+
+    public class Item : RealmObject
+    {
+        [PrimaryKey]
+        [MapTo("_id")]
+        public ObjectId Id { get; set; } = ObjectId.GenerateNewId();
+
+        [MapTo("owner_id")]
+        [Required]
+        public string OwnerId { get; set; }
+
+        [MapTo("summary")]
+        [Required]
+        public string Summary { get; set; }
+
+        [MapTo("isComplete")]
+        public bool IsComplete { get; set; }
+    }
+    //:snippet-end:
 }
+
