+++ source/sdk/dotnet/fundamentals/data-binding.txt
@@ -30,55 +30,53 @@ objects and collections update at the same time.
    forces loading the collection from memory, so the resulting collection is 
    no longer a live object.
 
-Data Binding a Realm Collection
--------------------------------
-In the following code example, we create a public ``IEnumerable`` of 
-Realm objects called "Items" and then populate that collection with all 
+Data Binding Examples
+---------------------
+The following code snippets show both the C# and XAML for 3 typical data 
+binding use cases. They all use the following Realm objects:
+
+.. literalinclude:: /examples/generated/dotnet/Models.snippet.binding-classes.cs
+   :language: csharp
+
+Binding to a Single Realm Object
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+In the following example, we create a public property of type Employee. Elsewhere 
+in our class, we query our realm for the Employee with an ``EmployeeId`` of "123":
+
+.. literalinclude:: /examples/generated/dotnet/MainPage.xaml.snippet.bind-one-cs.cs
+   :language: csharp
+
+In out XAML, we bind the Employee's ``Name`` property to a ``Label`` control's 
+``Text`` property:
+
+.. literalinclude:: /examples/generated/dotnet/MainPage.snippet.bind-one-xaml.xaml
+   :language: xml
+
+Binding to a Realm Collection
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+In this code example, we create a public ``IEnumerable`` of the ``Item``
+Realm object and then populate that collection with all 
 of the documents in the Items collection:
 
-.. code-block:: csharp
-   :copyable: false
-
-   ... 
-   
-   public IEnumerable<Item> Items { get; }
-   
-   ... 
-   
-   _itemsRealm = Realm.GetInstance(config);
-   Items = _itemsRealm.All<Item>();
+.. literalinclude:: /examples/generated/dotnet/MainPage.xaml.snippet.bind-list-cs.cs
+   :language: csharp
 
 In the associated XAML file, we bind a ``ListView`` to the **Items** ``IEnumerable`` 
 and then use the Item's property values to display information in the ListView:
 
-.. code-block:: xml
-   :copyable: false
-   :emphasize-lines: 1,11,13,14,17,20
+.. literalinclude:: /examples/generated/dotnet/MainPage.snippet.bind-list-xaml.xaml
+   :language: xml
 
-    <ListView x:Name="listItems" ItemsSource="{Binding Items}">
-        <ListView.ItemTemplate>
-            <DataTemplate>
-                <ViewCell>
-                    <Grid Padding="10,0,15,0" BackgroundColor="#ceF">
-                        <Grid.ColumnDefinitions>
-                            <ColumnDefinition Width="4*"/>
-                            <ColumnDefinition Width="1*"/>
-                            <ColumnDefinition Width="1*"/>
-                        </Grid.ColumnDefinitions>
-                        <Label Text="{Binding Summary}"/>
-                        <Switch x:Name="chkCompleted" 
-                           IsToggled="{Binding IsComplete}"/>
-                        <StackLayout Grid.Column="2" VerticalOptions="Center">
-                            <Image WidthRequest="30" x:Name="btnDelete" 
-                                    Source="trash.png">
-                                <Image.GestureRecognizers>
-                                    <TapGestureRecognizer Tapped="Delete_Clicked" 
-                                        CommandParameter="{Binding .}"/>
-                                </Image.GestureRecognizers>
-                            </Image>
-                        </StackLayout>
-                    </Grid>
-                </ViewCell>
-            </DataTemplate>
-        </ListView.ItemTemplate>
-    </ListView>
+Binding to an Embedded Collection
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+In this example, we bind to a collection of Items within the ``Items`` property 
+of a specific ``Employee``. 
+
+.. literalinclude:: /examples/generated/dotnet/MainPage.xaml.snippet.bind-embedded.cs
+   :language: csharp
+
+The XAML implementation is just like the example above:
+
+.. literalinclude:: /examples/generated/dotnet/MainPage.snippet.bind-embedded-xaml.xaml
+   :language: xml
+
